% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolandcompare.R
\name{poolandcompare}
\alias{poolandcompare}
\title{Pool counts and MOEs to compare two groups}
\usage{
poolandcompare(data, count_var, moe_var, total_var, moe_total = 0, group_var)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{count_var}{A character string representing the column name for the count variable.}

\item{moe_var}{A character string representing the column name for the margin of error (MOE) of the count variable.}

\item{total_var}{A character string representing the column name for the total count variable (denominator).}

\item{moe_total}{A character string representing the column name for the margin of error of the total count variable, or 0 if the total is a Census count. Default is 0.}

\item{group_var}{A character string for a variable that splits the data into two groups. Must be binary and code for 0 and 1.}
}
\value{
A list containing the percent values, z-score, p-value, and significance of the test.
}
\description{
This function pools a set of estimates and MOEs (such as from Census tabulations), splits them via a grouping variable, calculates percentages, and compares the percentages with significant test.
}
\examples{
library(tidycensus)
library(tidyverse)
library(pooltools)

# Download ACS data
# Include the variable for the number males and number of males in sales and service occupations
acs <- tidycensus::get_acs(geography = "county", 
                           variables = c("B01001_002", "B24022_028E"), 
                           output = "wide",
                           state = "MD")

# Get list of fips codes that border the coast
coastal_fips <- c("24003", "24005", "24009", "24011", "24015", "24017", 
                  "24019", "24025", "24029", "24035", "24039", "24037", 
                  "24041", "24045", "24047")

# Create a "coastal" variable for the ACS data
acs <- acs \%>\% 
  mutate(coastal = ifelse(GEOID \%in\% coastal_fips, 1, 0))

# Test whether counties that border the coast have a higher percentage of males in sales and service occupations
poolandcompare(
  data = acs,
  count_var = "B24022_028E", 
  moe_var = "B24022_028M", 
  total_var = "B01001_002E", 
  moe_total = "B01001_002M", 
  group_var = "coastal"
)

}
